move id and version to pkg
