+++ server/api/status.go
@@ -17,8 +17,8 @@ package api
 import (
 	"net/http"
 
+	"github.com/tikv/pd/pkg/versioninfo"
 	"github.com/tikv/pd/server"
-	"github.com/tikv/pd/server/versioninfo"
 	"github.com/unrolled/render"
 )
