+++ server/cluster/cluster_worker.go
@@ -24,10 +24,10 @@ import (
 	"github.com/tikv/pd/pkg/errs"
 	"github.com/tikv/pd/pkg/utils/logutil"
 	"github.com/tikv/pd/pkg/utils/typeutil"
+	"github.com/tikv/pd/pkg/versioninfo"
 	"github.com/tikv/pd/server/core"
 	"github.com/tikv/pd/server/schedule"
 	"github.com/tikv/pd/server/statistics/buckets"
-	"github.com/tikv/pd/server/versioninfo"
 	"go.uber.org/zap"
 )
