+++ tests/pdctl/helper.go
@@ -25,10 +25,10 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/pd/pkg/utils/typeutil"
+	"github.com/tikv/pd/pkg/versioninfo"
 	"github.com/tikv/pd/server"
 	"github.com/tikv/pd/server/api"
 	"github.com/tikv/pd/server/core"
-	"github.com/tikv/pd/server/versioninfo"
 	"github.com/tikv/pd/tests"
 )
